feat(cmd-serve): improving the flexibility of the `serve` command with port detection
The command now supports starting a PHP server with the ability to automatically detect an available port. The user is prompted to specify a port, and if the provided port is already in use, the command will suggest an alternative available port.

Users can also define custom scheme, host and port with optional arguments.

This enhancement provides a more user-friendly experience when running the **serve** command.

Asked by @eydun
Refs: #6